refactor: add send notification method to service
